requirements.txt
whiptail-dialogs
